update cmd/seeder with go subroutines and cli flags

+++ cmd/seeder/main.go
@@ -2,23 +2,35 @@ package main
 
 import (
 	"context"
+	"crypto/rand"
+	"flag"
 	"fmt"
+	"math/big"
 	"os"
+	"sync"
 	"time"
 
 	"github.com/jackc/pgx/v5/pgxpool"
 	"github.com/joho/godotenv"
 
-	"github.com/griggsjared/getsit/internal/entity"
 	"github.com/griggsjared/getsit/internal/repository"
 	"github.com/griggsjared/getsit/internal/service"
 )
 
 func main() {
 
-	godotenv.Load()
+	var tCount int
+	var wCount int
+	var fresh bool
+
+	flag.IntVar(&tCount, "n", 1000, "number of url entries to seed")
+	flag.IntVar(&wCount, "w", 4, "number of workers")
+	flag.BoolVar(&fresh, "f", false, "truncate url_entries table before seeding")
+	flag.Parse()
 
 	ctx := context.Background()
+
+	godotenv.Load()
 
 	dbUrl := os.Getenv("DATABASE_URL")
 	if dbUrl == "" {
@@ -37,19 +49,22 @@ func main() {
 
 	service := service.New(r)
 
-	db.Exec(ctx, "TRUNCATE url_entries")
+	if fresh {
+		db.Exec(ctx, "TRUNCATE url_entries")
+		fmt.Println("Truncated url_entries table")
+	}
 
-	SeedUrlEntries(ctx, 100000, service)
+	SeedUrlEntries(ctx, tCount, wCount, service)
 }
 
 // seed will generate a number of tokens and check for duplicates
-func SeedUrlEntries(ctx context.Context, tCount int, s *service.Service) {
+func SeedUrlEntries(ctx context.Context, tCount int, wCount int, s *service.Service) {
 
 	genCount := 0
 
 	timeStart := time.Now()
 
-	fmt.Println("Seeding", tCount, "url entries with random tokens")
+	fmt.Println("Seeding", tCount, "url entries with random tokens", "using", wCount, "workers")
 
 	fChan := make(chan bool)
 
@@ -60,18 +75,19 @@ func SeedUrlEntries(ctx context.Context, tCount int, s *service.Service) {
 			select {
 			case finished = <-fChan:
 				if finished {
+					genCount = tCount
 					percent = 100.0
 				}
 			default:
 				percent = float64(genCount) / float64(tCount) * 100
 			}
 			fmt.Printf("\rEntries Seeded: %d, Percent %.2f%%", genCount, percent)
-			time.Sleep(10 * time.Millisecond)
 			if finished {
 				fmt.Println("\nTime taken: ", time.Since(timeStart))
 				close(fChan)
 				break
 			}
+			time.Sleep(10 * time.Millisecond)
 		}
 	}()
 
@@ -80,21 +96,33 @@ func SeedUrlEntries(ctx context.Context, tCount int, s *service.Service) {
 		fChan <- true
 	}()
 
-	for i := 0; i < tCount; i++ {
+	wg := sync.WaitGroup{}
 
-		e, err := s.SaveUrl(ctx, &service.SaveUrlInput{
-			Url: "https://example.com/" + entity.NewUrlToken().String(),
-		})
+	perWorker := tCount / wCount
 
-		if err != nil {
-			fmt.Println(err)
-			break
-		}
+	for i := 0; i < wCount; i++ {
 
-		genCount++
+		wg.Add(1)
 
-		s.VisitUrlByToken(ctx, &service.VisitUrlByTokenInput{
-			Token: e.Token.String(),
-		})
+		go func() {
+			defer wg.Done()
+
+			for j := 0; j < perWorker; j++ {
+
+				rand, err := rand.Int(rand.Reader, big.NewInt(1000000000000000000))
+				if err != nil {
+					fmt.Println(err)
+					break
+				}
+
+				s.SaveUrl(ctx, &service.SaveUrlInput{
+					Url: "https://example.com/" + fmt.Sprintf("%d", rand),
+				})
+
+				genCount++
+			}
+		}()
 	}
+
+	wg.Wait()
 }
